:truck: Refactor : change import for domain directory
