+++ src/main/java/com/run_us/server/domains/running/exception/RunningErrorCode.java
@@ -7,7 +7,8 @@ public enum RunningErrorCode implements CustomResponseCode {
 
   RUNNING_NOT_FOUND("RE001", "Running not found", "Running not found", HttpStatus.NOT_FOUND),
   AGGREGATE_FAILED("RE4001", "Failed to Save Running Result", "Failed to Save Running Result",
-      HttpStatus.BAD_REQUEST),;
+      HttpStatus.BAD_REQUEST),
+  PERSONAL_RECORD_NOT_FOUND("REH4001", "Personal Record not found", "Personal Record not found", HttpStatus.BAD_REQUEST);
 
 
   private final String code;
