+++ src/main/java/com/run_us/server/domains/running/service/RunningLiveService.java
@@ -4,8 +4,8 @@ import static com.run_us.server.domains.running.domain.RunningConst.RUNNING_PREF
 import static com.run_us.server.domains.running.service.util.RunningKeyUtil.createLiveKey;
 
 import com.run_us.server.domains.running.domain.LocationData.RunnerPos;
-import com.run_us.server.domains.running.domain.ParticipantStatus;
-import com.run_us.server.domains.running.domain.Running;
+import com.run_us.server.domains.running.domain.running.ParticipantStatus;
+import com.run_us.server.domains.running.domain.running.Running;
 import com.run_us.server.domains.running.domain.RunningConst;
 import com.run_us.server.domains.running.exception.RunningErrorCode;
 import com.run_us.server.domains.running.exception.RunningException;
