+++ src/test/java/com/run_us/server/domains/running/RunningTest.java
@@ -2,9 +2,9 @@ package com.run_us.server.domains.running;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-import com.run_us.server.domains.running.domain.Running;
-import com.run_us.server.domains.running.domain.RunningConstraints;
-import com.run_us.server.domains.running.domain.RunningDescription;
+import com.run_us.server.domains.running.domain.running.Running;
+import com.run_us.server.domains.running.domain.running.RunningConstraints;
+import com.run_us.server.domains.running.domain.running.RunningDescription;
 import com.run_us.server.domains.user.domain.User;
 import com.run_us.server.domains.user.domain.UserFixtures;
 import com.run_us.server.global.util.PointGenerator;
